+++ src/Schema/AbstractTableSchema.php
@@ -152,9 +152,4 @@ abstract class AbstractTableSchema implements TableSchemaInterface
     {
         $this->foreignKeys[$id] = $to;
     }
-
-    public function compositeForeignKey(int $id, string $from, string $to): void
-    {
-        throw new NotSupportedException(static::class . ' does not support composite FK.');
-    }
 }